Add files via upload
An android application that will have two activities – Activity A and B.
Activity A will contain two EditText widgets for email and password and a button widget. When a user inputs email and password and click the button, whatever the user has inputted will be displayed on Activity B on a TextView widgets.